now we can create consumers and kinda consume then, however I have to find a way to read and then delete it

+++ Producer.h
@@ -21,7 +21,7 @@ private:
     float delayTime;
     int pid;
     int lifeSpan;
-    int startTime;
+    long startTime;
     int startingId;
 };
